requirements.txt
streamlit
langchain
openai
faiss-cpu
python-dotenv
PyMuPDF
tqdm

